+++ src/main/java/com/TimeVenture/repository/project/ProjecetRepository.java
@@ -3,6 +3,6 @@ package com.TimeVenture.repository.project;
 import com.TimeVenture.model.entity.project.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjecetRepository extends JpaRepository<Project, Long> {
+public interface ProjecetRepository extends JpaRepository<Project, Integer> {
 
 }
